move `test_entry.py`'s content into `test_entries.py`
